==> notebooks/2.0-db-creating-dataset.ipynb <==
import datetime
import gc
from itertools import product
import sys

sys.path.append('..')


import numpy as np
import pandas as pd
import scipy.stats as sps

import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.notebook import tqdm

import holidays

from src.utils.downcasting import downcast_dtypes

sns.set(font_scale=1.2)
%matplotlib inline

# Creating dataset

In this notebook I will create base dataset for training (without tuning for specific models).

items = pd.read_csv('../data/raw/items.csv')
item_categories = pd.read_csv('../data/raw/item_categories.csv')
shops = pd.read_csv('../data/raw/shops.csv')
sales_train = pd.read_csv('../data/raw/sales_train.csv')
test = pd.read_csv('../data/raw/test.csv')

## Cleaning

In this step I will clean dataset according to `1.0-db-EDA.ipynb`.

### Item categories

Now this section is empty, data was found clean.

### Shops

shops.head()

We found some typos in `shop_name`. We should replace shops with typos with shops without typos in `sales_train`, `test` (if it is the case).

map_typos = {0: 57, 1: 58, 10: 11, 39: 40}

Visualize mapping.

shops_list = shops.shop_name.to_list()
for from_id, to_id in map_typos.items():
    print(f"'{shops_list[from_id]}' ---> '{shops_list[to_id]}'")

Check datasets

for from_id, to_id in map_typos.items():
    print(f'Typos records with id={from_id} in train: '
          f'{sales_train.shop_id.isin([from_id]).sum()}')
    print(f'Corrected records wit id={to_id} in train: '
          f'{sales_train.shop_id.isin([to_id]).sum()}')

for from_id, to_id in map_typos.items():
    print(f'Typos records with id={from_id} in test: '
          f'{test.shop_id.isin([from_id]).sum()}')
    print(f'Corrected records wit id={to_id} in test: '
          f'{test.shop_id.isin([to_id]).sum()}')

As we can see, in test there is just one possible variant from two. We have to select one of them. We select corrected variants in all cases.

**If we've made a mistake here, we will check it in the future (run without mapping)**

sales_train.shop_id[sales_train.shop_id.isin(map_typos.keys())] = sales_train.shop_id[
    sales_train.shop_id.isin(map_typos.keys())
].map(map_typos)

test.shop_id[test.shop_id.isin(map_typos.keys())] = test.shop_id[
    test.shop_id.isin(map_typos.keys())
].map(map_typos)

Check online shops. Are they used in test dataset? If not we can just remove them to avoid dealing with Nans for some new features.

suspicious_shops = [9, 12, 55]
shops[shops.shop_id.isin(suspicious_shops)]

for shop in suspicious_shops:
    shop_name = shops[shops.shop_id == shop].shop_name.item()
    print(f"Records '{shop_name}' in train: "
          f"{(sales_train.shop_id == shop).sum()}")

for shop in suspicious_shops:
    shop_name = shops[shops.shop_id == shop].shop_name.item()
    print(f"Records '{shop_name}' in train: "
          f"{(test.shop_id == shop).sum()}")

In this case we can remove `Выездная торговля`.

sales_train.shape

sales_train = sales_train[sales_train.shop_id != 9]

### Items

items.head()

We found typos with `!`, `*`, `/`, `D`.

#### Correct `!`

first_character = items.item_name.apply(lambda x: x[0])
typos_list = items[first_character == '!'].item_name.to_list()
typos_list

There was no duplicates, thus just delete redundant `!`.

map_typos = {}
for typo in typos_list:
    map_typos[typo] = typo.strip('!').strip()
    
items.item_name[items.item_name.isin(map_typos.keys())] = items.item_name[
    items.item_name.isin(map_typos.keys())
].map(map_typos)

#### Correct `*`

typos_list = items[first_character == '*'].item_name.to_list()
typos_list

There was duplicate problem with `*МИХЕЙ И ДЖУМАНДЖИ  Сука любовь`. In all other cases we should just remove `*`.

map_typos = {}
for typo in typos_list:
    if typo[:3] != '*МИ':
        map_typos[typo] = typo.strip('*').strip()
    
items.item_name[items.item_name.isin(map_typos.keys())] = items.item_name[
    items.item_name.isin(map_typos.keys())
].map(map_typos)

Case with duplicate we should process separately. Check how many records with both typo and corrected.

items[items.item_name.str.contains('МИХЕЙ')]

map_typos = {12: 14690}

for from_id, to_id in map_typos.items():
    print(f'Typos records with id={from_id} in train: '
          f'{sales_train.item_id.isin([from_id]).sum()}')
    print(f'Corrected records wit id={to_id} in train: '
          f'{sales_train.item_id.isin([to_id]).sum()}')

for from_id, to_id in map_typos.items():
    print(f'Typos records with id={from_id} in test: '
          f'{test.item_id.isin([from_id]).sum()}')
    print(f'Corrected records wit id={to_id} in test: '
          f'{test.item_id.isin([to_id]).sum()}')

We can see, that we can make mapping without problems.

sales_train.item_id[sales_train.item_id.isin(map_typos.keys())] = sales_train.item_id[
    sales_train.item_id.isin(map_typos.keys())
].map(map_typos)

for from_id, to_id in map_typos.items():
    print(f'Typos records with id={from_id} in train: '
          f'{sales_train.item_id.isin([from_id]).sum()}')
    print(f'Corrected records wit id={to_id} in train: '
          f'{sales_train.item_id.isin([to_id]).sum()}')

#### Correct `/`

typos_list = items[first_character == '/'].item_name.to_list()
typos_list

There was no duplicates problem.

map_typos = {}
for typo in typos_list:
    map_typos[typo] = typo.strip('/')
    
items.item_name[items.item_name.isin(map_typos.keys())] = items.item_name[
    items.item_name.isin(map_typos.keys())
].map(map_typos)

#### Correct `D`

typos_list = items[items.item_name.str.endswith('    D')].item_name.to_list()
typos_list

There was no duplicates problem (it was already processed cases).

map_typos = {}
for typo in typos_list:
    map_typos[typo] = typo.strip('    D')
    
items.item_name[items.item_name.isin(map_typos.keys())] = items.item_name[
    items.item_name.isin(map_typos.keys())
].map(map_typos)

### Sales

sales_train.head()

In EDA we have found entry with negative price. 

sales_train[sales_train.item_price <= 1e-9]

Look at another entries with this `shop_id` and `item_id`.

sales_train[(sales_train.shop_id == 32) & (sales_train.item_id == 2973)]

It obviously looks like a mistake. Let's remove this row from the dataset.

sales_train = sales_train[sales_train.item_price > 0]

Let's delete too expensive items.

items_prices = sales_train.groupby('item_id').agg(
    {'item_price': ['median', 'min', 'max']}
).sort_values(('item_price', 'median'), ascending=False)
to_remove_items = items_prices.index[:3]
sales_train = sales_train[~sales_train.item_id.isin(to_remove_items)]

## Creating new features

In this step I will add new features according to `1.0-db-EDA.ipynb`.

### Item categories

item_categories.head()

#### Adding category and subcategory

We will split `item_category_name` by delimiter and treat parts as category and subcategory. If there is no delimiter, then it will be both: category and subcategory.

item_categories.item_category_name.to_list()[:10]

item_categories = item_categories.rename(columns={'item_category_name': 'item_full_category_name'})
item_categories['item_category_name'] = item_categories.item_full_category_name.apply(
    lambda x: x.split(' - ')[0]
)
item_categories['item_subcategory_name'] = item_categories.item_full_category_name.apply(
    lambda x: x.split(' - ')[-1]
)

#### Saving new dataset

item_categories.to_csv('../data/processed/item_categories.csv', index=False)

### Shops

shops.head()

#### Adding city

shops['city'] = shops.shop_name.apply(lambda x: x.strip('!').split()[0])

Delete city for `Выездная Торговля`.

Set city for `Интернет-магазин ЧС`, `Цифровой склад 1С-Онлайн` to `Online`.

shops.city[shops.shop_name == 'Выездная Торговля'] = None
shops.city[shops.shop_name.isin(['Интернет-магазин ЧС', 
                                 'Цифровой склад 1С-Онлайн'])] = 'Online'

Add number of residents in a city.

num_residents = pd.read_csv('../data/external/num_residents.csv', index_col=0)
num_residents.head()

shops['num_residents'] = shops.city.map(num_residents.num_residents)

#### Adding coordinates of a shop

To be added.

#### Adding shop_id in test_indicator

shops['shop_in_test'] = shops.shop_id.isin(test.shop_id)

#### Saving new dataset

shops.to_csv('../data/processed/shops.csv', index=False)

### Items

items.head()

#### Addint text features

Will be in 4.0-db-text-features.ipynb.

#### Adding item_id in test indicator

items.head()

items['item_in_test'] = items.item_id.isin(test.item_id)

#### Saving new dataset

items.to_csv('../data/processed/items.csv', index=False)

### Sales

sales_train.head()

#### Add revenue features

sales_train['revenue'] = sales_train['item_cnt_day'] * sales_train['item_price']

#### Adding date features

We also will process a part of test dataset here.

sales_train.date = pd.to_datetime(sales_train.date, format='%d.%m.%Y')
test_date = pd.DataFrame({'date': pd.date_range('2015-11-1', '2015-11-30', freq='D')})

for df in [sales_train, test_date]:
    df['day'] = df.date.dt.day
    df['month'] = df.date.dt.month
    df['year'] = df.date.dt.year
    df['weekday'] = df.date.dt.weekday

Add holiday features.

def days_since_holiday(datetimes):
    """Assign to each date days since last holiday."""
    ru_holidays = holidays.Russia()
    days_since_holiday = np.zeros(datetimes.shape[0])
    for i, current_datetime in tqdm(enumerate(datetimes.values), 
                                    total=datetimes.size):
        last_holiday = current_datetime
        while not last_holiday in ru_holidays:
            last_holiday = last_holiday - datetime.timedelta(days=1)
        days_since_holiday[i] = (current_datetime - last_holiday).days
    return days_since_holiday

russian_holidays = holidays.Russia()

for df in [sales_train, test_date]:
    df['is_holiday'] = df.date.apply(
        lambda x: x in russian_holidays
    )
    df['not_working_day'] = ((df['weekday'].isin([5, 6])) 
                             | (df['is_holiday']))
    df['days_since_holiday'] = days_since_holiday(
        df.date.dt.date
    ).astype(int)

#### Adding price features

We will device price on "even" and "non even" parts. Even part is a price, derived from first digit, non-even -- residual.

sales_train['item_non_even_price'] = sales_train.item_price.apply(
    lambda x: float(str(x)[1:])
)
sales_train['item_even_price'] = (sales_train.item_price 
                                  - sales_train['item_non_even_price'])
sales_train['fraction_non_even_price'] = (
    sales_train['item_non_even_price'] / sales_train['item_price']
)

#### Saving new dataset

sales_train.to_csv('../data/processed/sales_train.csv', index=False)

## Aggregating

### Grouping

Aggregate all features in one table by months. Just copy creating grid from `Programming_assignment_week_4`. 

# Create "grid" with columns
index_cols = ['shop_id', 'item_id', 'date_block_num']

# For every month we create a grid from all shops/items combinations from that month
grid = [] 
for block_num in sales_train['date_block_num'].unique():
    sales = sales_train[sales_train.date_block_num==block_num]
    grid.append(
        np.array(
            list(product(*[sales.shop_id.unique(), 
                           sales.item_id.unique(), [block_num]])), 
        )
    )

# Turn the grid into a dataframe
grid = pd.DataFrame(np.vstack(grid), columns = index_cols, dtype=np.int32)

# Groupby data to get shop-item-month aggregates
gb = sales_train.groupby(index_cols, as_index=False).agg(
    {'item_cnt_day': 'sum', 'revenue': 'sum'})
gb = gb.rename(columns={'item_cnt_day': 'target', 
                        'revenue': 'revenue_item_shop'})
train = pd.merge(grid, gb, how='left', on=index_cols).fillna(0)


# Same as above but with shop-month aggregates
gb = sales_train.groupby(['shop_id', 'date_block_num'], as_index=False).agg(
    {'item_cnt_day': 'sum', 'revenue': 'sum'}
)
gb = gb.rename(columns={'item_cnt_day': 'target_shop', 
                        'revenue': 'revenue_shop'})
train = pd.merge(train, gb, how='left', on=['shop_id', 'date_block_num']).fillna(0)


# Same as above but with item-month aggregates
gb = sales_train.groupby(['item_id', 'date_block_num'], as_index=False).agg(
    {'item_cnt_day': 'sum', 'revenue': 'sum'}
)
gb = gb.rename(columns={'item_cnt_day': 'target_item', 
                        'revenue': 'revenue_item'})
train = pd.merge(train, gb, how='left', on=['item_id', 'date_block_num']).fillna(0)


# Add num of num of days in a month, month, year
gb = sales_train.groupby('date_block_num', as_index=False).agg(
    {'date': 'nunique', 'item_cnt_day': 'sum', 
     'month': 'first', 'year': 'first'}
)
gb = gb.rename(columns={'date': 'num_days', 'item_cnt_day': 'target_month'})
train = pd.merge(train, gb, how='left', on='date_block_num').fillna(0)


# Add num of holidays, num of not working days, longest sequence without holidays
gb = sales_train[['date', 'date_block_num', 'is_holiday', 'not_working_day', 
                  'days_since_holiday']].drop_duplicates().groupby(
    'date_block_num', as_index=True
).agg(
    {'is_holiday': 'sum', 'not_working_day': 'sum', 
     'days_since_holiday': 'max'}
)
gb = gb.rename(columns={
    'is_holiday': 'num_holidays', 
    'not_working_day': 'num_not_working_days', 
    'days_since_holiday': 'longest_sequence_without_holidays'
})
train = pd.merge(train, gb, how='left', on='date_block_num').fillna(0)


# Add mean/std of price in a month (we will then make it lag feature)
gb = sales_train.groupby(['item_id', 'date_block_num'], as_index=False).agg(
    {'item_price': ['mean', 'std', 'nunique'], 
     'fraction_non_even_price': ['mean']}
)
gb.columns = [
    'item_id', 'date_block_num', 
    'price_mean', 'price_std', 'price_nunique', 'fraction_non_even_mean',
]
train = pd.merge(train, gb, how='left', on=['item_id', 'date_block_num']).fillna(0)

# Add change of price in a month compare to global mean
gb = sales_train.groupby('item_id').agg({'item_price': 'mean'}).item_price
train['relative_delta_price_mean'] = (
    (train['price_mean'] - train['item_id'].map(gb)) 
    / train['item_id'].map(gb)
)

Reduce memory usage.

train = downcast_dtypes(train)

del grid, gb
gc.collect();

Features with averaging. We will add:
* average target for month
* average target for month and city

These features will be lag features lately.

gb = train.groupby('date_block_num').agg({'target': 'mean'})
train['mean_target_month'] = train.date_block_num.map(gb.target)

shop_to_city = pd.Series(data=shops.city, index=shops.shop_id)
train['city'] = train['shop_id'].map(shop_to_city)
gb = train.groupby(['date_block_num', 'city']).agg({'target': 'mean'})
gb = train.groupby(['date_block_num', 'city']).agg({'target': 'mean'})
gb = gb.rename(columns={'target': 'mean_target_month_city'})
train = pd.merge(train, gb, how='left', on=['date_block_num', 'city']).fillna(0)
train.drop(columns=['city'], inplace=True)

train = downcast_dtypes(train)

del gb
gc.collect();

### Adding lags features

Before calculating lag features let's join train and test dataset. We add redundant `target, target_shop`, `target_item`, `price_xxx`, `fraction_xxx`, `revenue_xxx` just to process united dataset.

test = test.drop(columns=['ID'])
test['date_block_num'] = train.date_block_num.max() + 1

test['num_days'] = test_date.shape[0]
test['month'] = test_date.month[0]
test['year'] = test_date.year[0]
test['num_holidays'] = test_date.is_holiday.sum()
test['num_not_working_days'] = test_date.not_working_day.sum()
test['longest_sequence_without_holidays'] = test_date.days_since_holiday.max()

for column in train.columns.difference(test.columns):
    test[column] = 0

all_data = pd.concat([train, test])
del train
gc.collect()

After creating a grid, we can calculate lag features. We will use lags from [1, 2, 3, 6, 12] months ago.

# List of columns that we will use to create lags
cols_to_rename = list(all_data.columns.difference(
    index_cols 
    + ['month', 'year', 'num_days', 'num_holidays', 
       'num_not_working_days', 'longest_sequence_without_holidays']
))

shift_range = [1, 2, 3, 6, 12]

for month_shift in tqdm(shift_range):
    train_shift = all_data[index_cols + cols_to_rename].copy()
    
    train_shift['date_block_num'] = train_shift['date_block_num'] + month_shift
    
    foo = lambda x: '{}_lag_{}'.format(x, month_shift) if x in cols_to_rename else x
    train_shift = train_shift.rename(columns=foo)

    all_data = pd.merge(all_data, train_shift, on=index_cols, how='left').fillna(0)
    all_data = downcast_dtypes(all_data)
    
    del train_shift
    gc.collect()

Remove earliest months with no info about lags, remove target columns.

# Don't use old data from year 2013
all_data = all_data[all_data['date_block_num'] >= 12] 
target = all_data.target

# List of all lagged features
fit_cols = [
    col for col in all_data.columns 
    if col.split('_')[-1] in [str(item) for item in shift_range]
]
# We will drop these at fitting stage
to_drop_cols = list(
        set(list(all_data.columns)) 
        - (set(fit_cols)|set(all_data.columns.difference(cols_to_rename)))
)
all_data = all_data.drop(to_drop_cols, axis=1)

### Merging with previous datasets

Join items with items_categories.

items_with_categories = pd.merge(
    items, 
    item_categories, 
    how='left', on='item_category_id'
).drop(columns=['item_category_id'])

Join train with items.

all_data = pd.merge(
    all_data,
    items_with_categories,
    how='left', on='item_id'
)
gc.collect();

Join train with shops.

all_data = pd.merge(
    all_data,
    shops,
    how='left', on='shop_id'
)
gc.collect();

#### Saving new datasets

all_data = downcast_dtypes(all_data)
all_data.info()

test_size = test.shape[0]
train = all_data.iloc[:-test_size]
target = target.iloc[:-test_size]
train['target'] = target.values
test = all_data.iloc[-test_size:]

train.reset_index(inplace=True)
test.reset_index(inplace=True)

As we know from `1.0-db-EDA.ipynb` a lot of pairs (item, shop) in test don't present in train. During our lar creating we added this pairs to train with a lot of zeros after filling NaNs. It can unintentionally mislead our model. Let's delete this pairs from train.

sales_train_pairs = set(
    map(
        lambda x: f'{x[0]}_{x[1]}', 
        zip(sales_train.item_id, sales_train.shop_id)
    )
)

original_test = pd.read_csv('../data/raw/test.csv')
original_test_pairs = set(
    map(
        lambda x: f'{x[0]}_{x[1]}', 
        zip(original_test.item_id, original_test.shop_id)
    )
)

problem_pairs = original_test_pairs.difference(sales_train_pairs)

train_pairs = pd.Series(
    map(
        lambda x: f'{x[0]}_{x[1]}', 
        zip(train.item_id, train.shop_id)
    )
)

train.shape

train = train[~train_pairs.isin(problem_pairs)]

For increase the speed of reading train dataset, let's save it in feather format.

train.to_feather('../data/processed/train.ftr')
test.to_feather('../data/processed/test.ftr')
