==> notebooks/3.0-db-baseline.ipynb <==
import numpy as np
import pandas as pd
import scipy.stats as sps

import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.notebook import tqdm

sns.set(font_scale=1.2)
%matplotlib inline

# Baseline

In this notebook I will provide a baseline solution according to advice 1, 2 in the course.

Advice 1:
> Competition data is rather challenging, so the sooner you get yourself familiar with it - the better. You can start with submitting sample_submission.csv from "Data" page on Kaggle and try submitting different constants.

Advice 2:

> A good exercise is to reproduce previous_value_benchmark. As the name suggest - in this benchmark for the each shop/item pair our predictions are just monthly sales from the previous month, i.e. October 2015.

> The most important step at reproducing this score is correctly aggregating daily data and constructing monthly sales data frame. You need to get [lagged](https://en.wikipedia.org/wiki/Lag_operator) values, fill NaNs with zeros and clip the values into [0,20] range. If you do it correctly, you'll get precisely 1.16777 on the public leaderboard.

> Generating features like this is a necessary basis for more complex models. Also, if you decide to fit some model, don't forget to clip the target into [0,20] range, it makes a big difference.

## Sample submission

According to advice 1 just send sample submission.

submission = pd.read_csv('../data/raw/sample_submission.csv')

submission.head()

Check values of `item_cnt_month`.

submission.item_cnt_month.unique()

submission.to_csv('../models/constants/0.5.csv', index=False)

As we can see, we predict just one constant value. Send it to the system using `kaggle` utility.

!kaggle competitions submit competitive-data-science-predict-future-sales -f ../data/raw/sample_submission.csv -m "Sample submission"

!kaggle competitions submissions competitive-data-science-predict-future-sales

We got $1.23646$ on the liderboard.

## Different constants

In this section we will try another constants and find the optima.

submission.item_cnt_month = 1.0
submission.to_csv('../models/constants/1.0.csv', index=False)

!kaggle competitions submit competitive-data-science-predict-future-sales -f ../models/constants/1.0.csv -m "Constant prediction 1.0"

!kaggle competitions submissions competitive-data-science-predict-future-sales

We got $1.41241$ on the liderboard.

submission.item_cnt_month = 0.0
submission.to_csv('../models/constants/0.0.csv', index=False)

!kaggle competitions submit competitive-data-science-predict-future-sales -f ../models/constants/0.0.csv -m "Constant prediction 0.0"

!kaggle competitions submissions competitive-data-science-predict-future-sales

We got $1.25011$ on the liderboard.

Now we can easily calculate the best constant for MSE. Assume that we have predictions $y_1, y_2$ and their scores $\alpha, \beta$. It is known, that we can represent MSE in that case as

$$
MSE = (y - \overline{y})^2 + m, 
$$
ther $m$ is the lowest possible constant error.

If we have two submissions, then we can get the system of equations. By substracting it we can get a formula:

$$
\overline{y} = \frac{1}{2} \left( y_1 + y_2 - \frac{\alpha - \beta}{y_1 - y_2} \right)
$$

If you take into account, that LB tells us RMSE, you get: $\overline{y} = 0.2839$.

Let's submit this prediction.

submission.item_cnt_month = 0.2839
submission.to_csv('../models/constants/0.2839.csv', index=False)

!kaggle competitions submit competitive-data-science-predict-future-sales -f ../models/constants/0.2839.csv -m "Constant prediction 0.2839"

!kaggle competitions submissions competitive-data-science-predict-future-sales

We got $1.21743$ on the liderboard.

## Previous value benchmark

According to advice 2 for each `item_id`, `shop_id` send prediction for the previous month and clip the value within [0, 20].

test = pd.read_hdf('../data/processed/test.h5', 'test')
test.head()

submission.head()

submission['item_cnt_month'] = np.clip(test['target_lag_1'].values, 0, 20)

submission.to_csv('../models/previous_value/submission.csv', index=False)

As we can see, we predict just one constant value. Send it to the system using `kaggle` utility.

!kaggle competitions submit competitive-data-science-predict-future-sales -f ../models/previous_value/submission.csv -m "Previous value benchmark"

!kaggle competitions submissions competitive-data-science-predict-future-sales

We got $1.16777$ on the liderboard as was predicted by advice.
