==> notebooks/4.0-db-text-features.ipynb <==
import sys
import re

sys.path.append('..')

import numpy as np
import pandas as pd
import scipy.stats as sps

import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.notebook import tqdm

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import TruncatedSVD

from nltk.stem.snowball import SnowballStemmer

from src.utils.downcasting import downcast_dtypes

sns.set(font_scale=1.2)
%matplotlib inline

# I have found, that this setting is the best
DO_CLEANING = True
DO_STEMMING = True
RES_DIM = 20

random_state = 42

def correct_name(x):
    """Corrects string.
    
    :param x: string
    
    :returns: corrected string
    """
    x = x.lower() # all letters lower case
    x = x.partition('[')[0] # partition by square brackets
    x = x.partition('(')[0] # partition by curly brackets
    x = re.sub('[^A-Za-z0-9А-Яа-я]+', ' ', x) # remove special characters
    x = x.replace('  ', ' ') # replace double spaces with single spaces
    x = x.strip() # remove leading and trailing white space
    return x

# Text features

As we found in 1.0-db-EDA.ipynb we can produce text features for items. It will be helpful, because there are a lot of items and we want to treat similar similarly.

## TFiDF + TruncatedSVD

First approach is to use TFiDF + TruncatedSVD. We will use $50$ features and in the future we will be able to select just part of them.

It will be difficult to find optimum hyperparameters because operation of learning TFiDF separately and merging it with train seems too complicated.

May be will be better to find reasonable parameters for some model by hand (for example for random forest) and fix them.

items = pd.read_csv('../data/processed/items.csv')
items.head()

names = items.item_name

Let's clean item names.

if DO_CLEANING:
    names = names.apply(correct_name)

Let's stem item names.

if DO_STEMMING:
    stemmer = SnowballStemmer('russian')
    names = names.apply(lambda x: stemmer.stem(x))

Make TFiDF transformation.

transformer = TfidfVectorizer(ngram_range=(1, 1), max_df=1.0)
tfidf_data = transformer.fit_transform(names)

tfidf_data.shape

svd = TruncatedSVD(n_components=RES_DIM, random_state=random_state)
truncated_data = svd.fit_transform(tfidf_data)

truncated_data.shape

Save result on disc.

columns = [f'item_name_tfidf_truncated-svd_{i+1}' for i in range(RES_DIM)]
df = pd.DataFrame(truncated_data, columns=columns)

df = downcast_dtypes(df, to_bytes=16)
df.to_feather('../data/processed/text/tfidf_truncated-svd.ftr')
